+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # @langtrase/typescript-sdk
 
+## 1.3.3
+- Fix for function calling response tracing
+- Fix for streaming prompt calculation
+
 ## 1.3.2
 
 - Adds 'langtrace.sdk.name' to the trace attributes